pyproject.toml
[tool.poetry]
name = "tjjobs"
version = "0.1.0"
description = "TJ's Scheduled Jobs"
authors = ["TJ Palanca <mail@tjpalanca.com>"]
license = "AGPL-3.0"
repository = "https://github.com/tjpalanca/tjjobs"
keywords = ["dagster", "jobs"]

[tool.poetry.dependencies]
python = "^3.10"
dagster = "^0.15.6"
dagit = "^0.15.6"
requests = "^2.28.1"
tjhome = {git = "https://github.com/tjpalanca/tjhome.git", rev = "b98629a"}
dagster-postgres = "^0.15.6"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
flake8 = "^4.0.1"
black = "^22.6.0"
ipython = "^8.4.0"
bandit = "^1.7.4"
python-on-whales = "^0.49.0"
poetry-core = "^1.0.8"
requests = "^2.28.1"

[tool.poetry.scripts]
image_build = "scripts.image:build"
image_shell = "scripts.image:shell"
image_push = "scripts.image:push"
image_run = "scripts.image:run"
tjjobs = "tjjobs.__main__:main"
deploy = "scripts.deploy:deploy"
build_and_deploy = "scripts.deploy:build_and_deploy"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
